feat: add TildaRegistryEntity to models project
