+++ src/Dan.Plugin.Tilda/Models/TildaEnums.cs
@@ -4,27 +4,27 @@ using Newtonsoft.Json.Converters;
 
 namespace Dan.Plugin.Tilda.Models
 {
-    [JsonConverter(typeof(StringEnumConverter))]
-    public enum OperationStatus
-    {
-        [EnumMember(Value = "ikkeAngitt")]
-        Blank = 0,
-
-        [EnumMember(Value = "konkurs")]
-        Konkurs = 1,
-
-        [EnumMember(Value = "underAvvikling")]
-        UnderAvvikling = 2,
-
-        [EnumMember(Value = "underTvangsavviklingEllerTvangsopploesning")]
-        UnderTvangsavviklingEllerTvangsopplosning = 3,
-
-        [EnumMember(Value = "ok")]
-        OK = 4,
-
-        [EnumMember(Value = "slettet")]
-        Slettet = 5
-    }
+    // [JsonConverter(typeof(StringEnumConverter))]
+    // public enum OperationStatus
+    // {
+    //     [EnumMember(Value = "ikkeAngitt")]
+    //     Blank = 0,
+    //
+    //     [EnumMember(Value = "konkurs")]
+    //     Konkurs = 1,
+    //
+    //     [EnumMember(Value = "underAvvikling")]
+    //     UnderAvvikling = 2,
+    //
+    //     [EnumMember(Value = "underTvangsavviklingEllerTvangsopploesning")]
+    //     UnderTvangsavviklingEllerTvangsopplosning = 3,
+    //
+    //     [EnumMember(Value = "ok")]
+    //     OK = 4,
+    //
+    //     [EnumMember(Value = "slettet")]
+    //     Slettet = 5
+    // }
 
     [JsonConverter(typeof(StringEnumConverter))]
     public enum ControlTopic
